Add files via upload
This was an attempt at implementing what I learned from Coursera's 'Introduction to TensorFlow for Artificial Intelligence, Machine Learning, and Deep Learning'. I used a dataset from Kaggle and reused some of my own code from the assignments for the course. 
The data set was: https://www.kaggle.com/datasets/jonathanoheix/face-expression-recognition-dataset/data
The accuracy is very bad but it's a first attempt, so it wasn't too bad.